Filter Feature refactored to use Redux store (#116)
Merged PR-116
