+++ app-web/src/store/actions/actionTypes.js
@@ -6,3 +6,6 @@ export const LOAD_SIPHON_NODES = 'LOAD_SIPHON_NODES';
 export const LOAD_SIPHON_NODES_SUCCESS = 'LOAD_SIPHON_NODES_SUCCESS';
 export const LOAD_SIPHON_NODES_FAIL = 'LOAD_SIPHON_NODES_FAIL';
 export const FILTER_SIPHON_NODES = 'FILTER_SIPHON_NODES';
+// UI ACTIONS
+export const SET_SELECTED_FILTER_OPTION = 'SET_SELECTED_FILTER_OPTION';
+export const TOGGLE_MAIN_NAVIGATION = 'TOGGLE_MAIN_NAVIGATION';
