+++ app-web/src/store/actions/actions.js
@@ -1,2 +1,3 @@
 export { authenticateSuccess, authenticateFailed } from './authActions';
 export { loadSiphonNodes, filterSiphonNodes } from './siphonActions';
+export { setSelectedFilterOption, toggleMainNavigation } from './uiActions';
