+++ app-web/src/store/reducers/ui.js
@@ -18,20 +18,16 @@ Created by Patrick Simonian
 import * as actionTypes from '../actions/actionTypes';
 
 const initialState = {
-  isAuthenticated: false,
-  accessToken: null,
-  idToken: null,
-  error: false,
-  messages: [],
+  selectedFilterOption: null,
+  mainNavigationToggled: false,
 };
 
 const reducer = (state = initialState, action) => {
   switch (action.type) {
-    case actionTypes.AUTHENTICATE_SUCCESS:
-      console.log('called');
-      return { ...state, isAuthenticated: true };
-    case actionTypes.AUTHENTICATE_FAILED:
-      return { ...state, isAuthenticated: false };
+    case actionTypes.SET_SELECTED_FILTER_OPTION:
+      return { ...state, selectedFilterOption: action.payload };
+    case actionTypes.TOGGLE_MAIN_NAVIGATION:
+      return { ...state, mainNavigationToggled: action.payload.toggled };
     default:
       return state;
   }
